+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 * Okanshi.Owin now has an option for which timer to use when monitoring.
 * `SlaTimer` has been added. The timer makes it easy to track SLA-breaches.
+* `CounterZeroFilter`, `GaugeZeroFilter` and `TimerZeroFilter` has been added. They wrap monitors and only returns values to the poller when changes have been registered.
 * `MinMaxAvgGauge` has been added. It allows you to detect extreme values that would otherwise disappear in an average calculation.
 * `OkanshiMonitor.DefaultStep` has been deleted since it was forgotten in an earlier clean up process.
 * Unnecessary calculation has been removed from `Counter`, and locks are no longer used in the counter, instead atomics are used.